Remove observability package
