+++ legacy/deploy/observability/grafana-datasources.yaml
@@ -12,7 +12,7 @@ datasources:
       maxLines: 5000
       derivedFields:
         # Click-through from a log line to the trace it belongs to (logfire /
-        # OTel emit a trace_id; the observability package puts it in
+        # OTel emit a trace_id; legacy logging puts it in
         # obs_context when present).
         - name: trace_id
           matcherRegex: "\"trace_id\"\\s*:\\s*\"([a-f0-9]+)\""
