+++ legacy/observability/integrations/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy-local observability integrations."""
