+++ potpie/integrations/integrations/adapters/outbound/oauth/linear_oauth.py
@@ -9,7 +9,7 @@ import httpx
 import urllib.parse
 import time
 import hashlib
-from observability import get_logger
+from integrations.logging import get_logger
 from integrations import hash_user_id
 
 logger = get_logger(__name__)
